rag-api/requirements.txt
fastapi
uvicorn
pymilvus
openai
pydantic
python-dotenv
python-multipart
uuid
